architecture: extract setInterval loops from BuildingGameClient into useBuildingGameLoop
BuildingGameClient.tsx had two bare useEffect/setInterval game-loop pairs
directly in the component body, mixing rendering concerns with timing logic.

Extracted both into a dedicated useBuildingGameLoop hook:
  - Particle tick (50 ms): advances particle positions when any particles exist
  - Animation tick (100 ms): rotates blocks 2° each tick when animationMode is on

BuildingGameClient now calls useBuildingGameLoop() and removes the direct
useEffect import.

Closes #604
